+++ frontend/plugins/tourism_ui/src/modules/tms/branch-detail/dashboard/elements/index.ts
@@ -4,6 +4,10 @@ export { ElementEditSheet } from './_components/ElementEditSheet';
 export { ElementCommandBar } from './_components/ElementCommandBar';
 export { ElementFilter } from './_components/ElementFilter';
 export { elementColumns } from './_components/ElementColumns';
+export {
+  ElementMoreColumn,
+  elementMoreColumn,
+} from './_components/ElementMoreCell';
 
 export { useElements } from './hooks/useElements';
 export { useCreateElement } from './hooks/useCreateElement';