+++ frontend/plugins/tourism_ui/src/modules/tms/branch-detail/dashboard/itinerary/types/itinerary.ts
@@ -5,4 +5,31 @@ export interface IItinerary {
   duration?: number;
   createdAt?: string;
   modifiedAt?: string;
+  color?: string;
+  images?: string[];
+  content?: string;
+  totalCost?: number;
+  guideCost?: number;
+  driverCost?: number;
+  foodCost?: number;
+  gasCost?: number;
+  personCost?: Record<string, number>;
+  guideCostExtra?: number;
+  groupDays?: Array<{
+    __typename?: string;
+    day?: number;
+    title?: string;
+    content?: string;
+    elements?: Array<{
+      __typename?: string;
+      elementId?: string;
+      orderOfDay?: number;
+    }>;
+    elementsQuick?: Array<{
+      __typename?: string;
+      elementId?: string;
+      orderOfDay?: number;
+    }>;
+    images?: string[];
+  }>;
 }