+++ frontend/plugins/tourism_ui/src/pages/tms/IndexPage.tsx
@@ -14,7 +14,7 @@ export const IndexPage = () => {
             <Breadcrumb.List className="gap-1">
               <Breadcrumb.Item>
                 <Button variant="ghost" asChild>
-                  <Link to="/tms">
+                  <Link to="/tourism/tms">
                     <IconBox />
                     Tour management system
                   </Link>